begin making more stashes of functions..
